requirements.txt
fastapi
uvicorn
httpx
python-dotenv

openai

pinecone-client
pinecone
langchain
langchain-community
langchain-text-splitters

pypdf

supabase

tiktoken

pydantic

sentence-transformers